Make proxy_device_resolve_by_minor function to be a O(1) algorithm
Initially it was O(n) algorithm, but we don't really need to run in
a loop to find the corresponding proxy device structure at all,
as we initialize each structure one after another and increment the
minor number by 1, essentially mapping the index number to the minor
number for all of ufedm_proxy%d devices.

+++ src/proxy_device/class.h
@@ -11,6 +11,6 @@
 int proxy_device_class_init(size_t dev_count);
 void proxy_device_class_exit(void);
 
-struct ufedm_proxy_device *proxy_device_resolve_by_minor(int minor);
+struct ufedm_proxy_device *proxy_device_resolve_by_minor(size_t minor);
 
 #endif
